Added moving blocks
Made the 'UpDown' class, a derivative of the 'MovementType'. This means we can create a DynamicBlock that can move up and down, this has been added to the main function as a demo.

This breaks the player collision quite considerably, as can be seen if you actually run the code. I decided to commit anyway as the purpose was to get moving blocks working and that does seem to work. I will start work on the player collision immediatly.