Added changing layouts
The selected layout now changes every time the player scores a point. This change might be hard to see at the moment because it randomly picks from 2 choices, so it could just pick the same layout over and over. Nonetheless, this is effectively the major gameplay element, so it's good to see it added.

All this really leaves to do mechanically is add the clock, death, and sort out the DynamicBlocks and MovementType, then it's just a matter of adding a load of layouts and polishing up the animations and the like.