Added DeathBlock
It's a StaticBlock that flashes to warn the player that it's super spooky, and that's about it. I thought about modifying the StaticBlock class to allow a switch to make it flash or not, but I thought it was unnessecary overhead to add a useless timer to every StaticBlock. If someone more experienced at OOP would love to comment, please do.

Also, this creates a permenant change to how the game looks, so screenshot update!