Fixed 'sticky' collision
Huzzah! This should be most of the collision problems delt with. It's not perfect and there are a couple of loose ends that need patching up, but I think this is fine to merge. I also think the method I've used is a tad hacky so I'll revisit it once I've thought about it, but for now we're good.