Well this is odd
In a vague attempt at debugging I tried commenting out a large section, and it seems to work perfectly? At least on my laptop where collisions worked before this branch anyway, will have to test on netbook (which is the main purpose of this push). This is very odd but if it works it works and I can merge soonish.